emco-sales/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
colorthief==0.2.1
Pillow==10.2.0
